requirements.txt
# Core Dependencies
requests>=2.31.0          # HTTP requests for APIs
PyYAML>=6.0               # For config.yml parsing
python-dotenv>=1.0.0        # Environment variable management

# Git Integrations
PyGithub>=1.59.0           # GitHub API
python-gitlab>=4.0.0       # GitLab API
requests-oauthlib>=1.3.0   # OAuth for Bitbucket

# Code Analysis Tools
flake8>=6.0.0              # Style + lint checks
radon>=6.0.0              # Complexity analysis
bandit>=1.7.0             # Security analysis
safety>=2.3.0             # Dependency vulnerability scanning
mypy>=1.5.0               # Type checking
black>=23.0.0              # Code formatting
isort>=5.12.0              # Import sorting

# AI Integration
openai>=1.0.0              # AI-driven feedback
anthropic>=0.7.0           # Alternative AI provider

# Web Framework
fastapi>=0.104.0          # For web demo dashboard
uvicorn[standard]>=0.24.0  # ASGI server
jinja2>=3.1.0             # Template engine
python-multipart>=0.0.6   # Form data handling
streamlit>=1.28.0         # Simple web interface
plotly>=5.17.0            # Interactive charts
pandas>=2.1.0             # Data manipulation

# CLI Interface
rich>=13.0.0               # Rich terminal formatting
click>=8.1.0               # CLI framework
typer>=0.9.0               # Modern CLI framework

# Testing
pytest>=7.4.0             # Testing framework
pytest-cov>=4.1.0         # Coverage reporting
pytest-mock>=3.11.0       # Mocking utilities
pytest-asyncio>=0.21.0    # Async testing
httpx>=0.25.0              # HTTP client for testing

# Development Tools
pre-commit>=3.5.0         # Git hooks
black>=23.0.0              # Code formatting
flake8>=6.0.0              # Linting
mypy>=1.5.0                # Type checking
isort>=5.12.0              # Import sorting

# Optional Enhancements
redis>=5.0.0               # Caching (optional)
celery>=5.3.0              # Task queue (optional)
prometheus-client>=0.17.0   # Metrics (optional)
